text field: extract message to own reusable composable
