+++ ui/src/main/java/kiwi/orbit/compose/ui/controls/TextField.kt
@@ -1,26 +1,16 @@
 package kiwi.orbit.compose.ui.controls
 
-import androidx.compose.animation.AnimatedContent
-import androidx.compose.animation.ExperimentalAnimationApi
-import androidx.compose.animation.SizeTransform
 import androidx.compose.animation.animateColor
 import androidx.compose.animation.core.tween
 import androidx.compose.animation.core.updateTransition
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.animation.slideInVertically
-import androidx.compose.animation.slideOutVertically
-import androidx.compose.animation.with
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.interaction.collectIsFocusedAsState
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
@@ -32,14 +22,13 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.SolidColor
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
-import kiwi.orbit.compose.icons.Icons
 import kiwi.orbit.compose.ui.OrbitTheme
 import kiwi.orbit.compose.ui.controls.field.FieldContent
+import kiwi.orbit.compose.ui.controls.field.FieldMessage
 import kiwi.orbit.compose.ui.foundation.LocalTextStyle
 import kiwi.orbit.compose.ui.foundation.ProvideMergedTextStyle
 
@@ -154,63 +143,7 @@ public fun TextField(
                 }
             )
 
-            Message(error, info, textStyle)
-        }
-    }
-}
-
-@OptIn(ExperimentalAnimationApi::class)
-@Composable
-private fun Message(
-    error: @Composable (() -> Unit)? = null,
-    info: @Composable (() -> Unit)? = null,
-    textStyle: TextStyle
-) {
-    val state = when {
-        error != null -> Message.Error(error)
-        info != null -> Message.Info(info)
-        else -> null
-    }
-    AnimatedContent(
-        targetState = state,
-        transitionSpec = {
-            if (targetState == null || initialState == null) {
-                val enter = slideInVertically(animationSpec = tween(AnimationDuration)) +
-                    fadeIn(animationSpec = tween(AnimationDuration))
-                val exit = slideOutVertically(animationSpec = tween(AnimationDuration)) +
-                    fadeOut(animationSpec = tween(AnimationDuration))
-                val size = SizeTransform(clip = false) { _, _ -> tween(AnimationDuration) }
-                enter with exit using size
-            } else {
-                val enter = fadeIn(animationSpec = tween(AnimationDuration))
-                val exit = fadeOut(animationSpec = tween(AnimationDuration))
-                val size = SizeTransform(clip = false) { _, _ -> tween(AnimationDuration) }
-                enter with exit using size
-            }
-        },
-    ) { message ->
-        if (message != null) {
-            Row(Modifier.padding(top = 6.dp)) {
-                val icon = when (message) {
-                    is Message.Error -> Icons.AlertCircle
-                    is Message.Info -> Icons.InformationCircle
-                }
-                val tintColor = when (message) {
-                    is Message.Error -> OrbitTheme.colors.critical.main
-                    is Message.Info -> OrbitTheme.colors.interactive.main
-                }
-                Icon(
-                    icon,
-                    contentDescription = null,
-                    modifier = Modifier
-                        .padding(top = 2.dp, end = 4.dp)
-                        .size(16.dp),
-                    tint = tintColor,
-                )
-                ProvideMergedTextStyle(textStyle.copy(color = tintColor)) {
-                    message.content.invoke()
-                }
-            }
+            FieldMessage(error, info, textStyle)
         }
     }
 }
@@ -220,13 +153,6 @@ private enum class InputState {
     NormalError,
     Focused,
     FocusedError,
-}
-
-private sealed class Message(
-    open val content: @Composable (() -> Unit)
-) {
-    data class Error(override val content: @Composable () -> Unit) : Message(content)
-    data class Info(override val content: @Composable () -> Unit) : Message(content)
 }
 
 private const val AnimationDuration = 150
